requirements.txt
# requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
coloredlogs==15.0.1
greenlet==3.0.3
humanfriendly==10.0
idna==3.8
pyreadline3==3.4.1
pyTelegramBotAPI==4.22.1
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.2
psycopg2==2.9.9
